Move type_descriptor implementation inside trait definition
