+++ crates/mirror-mirror-macros/src/derive_reflect/enum_.rs
@@ -267,8 +267,6 @@ fn expand_reflect(
         quote! {
             fn type_descriptor(&self) -> Cow<'static, TypeDescriptor> {
                 impl #impl_generics Typed for #ident #type_generics #where_clause {
-                    fn_type_descriptor!();
-
                     fn build(graph: &mut TypeGraph) -> NodeId {
                         let variants = &[#(#code_for_variants),*];
                         graph.get_or_build_node_with::<Self, _>(|graph| {
