+++ crates/mirror-mirror/src/lib.rs
@@ -291,52 +291,6 @@ macro_rules! trivial_reflect_methods {
     };
 }
 
-#[cfg(feature = "std")]
-#[macro_export]
-macro_rules! fn_type_descriptor {
-    () => {
-        fn type_descriptor() -> $crate::__private::Cow<'static, $crate::type_info::TypeDescriptor> {
-            use std::collections::HashMap;
-            use std::sync::RwLock;
-            use $crate::__private::*;
-
-            // a map required for generic types to have different type descriptors such as
-            // `Vec<i32>` and `Vec<bool>`
-            static INFO: OnceBox<RwLock<HashMap<TypeId, &'static TypeDescriptor>>> = OnceBox::new();
-
-            let type_id = TypeId::of::<Self>();
-
-            let lock = INFO.get_or_init(Box::default);
-            if let Some(info) = lock.read().unwrap().get(&type_id) {
-                return Cow::Borrowed(info);
-            }
-
-            let mut map = lock.write().unwrap();
-            let info = map.entry(type_id).or_insert_with(|| {
-                let mut graph = TypeGraph::default();
-                let id = Self::build(&mut graph);
-                let info = TypeDescriptor::__private_new(id, graph);
-                Box::leak(Box::new(info))
-            });
-            Cow::Borrowed(*info)
-        }
-    };
-}
-
-#[cfg(not(feature = "std"))]
-#[macro_export]
-macro_rules! fn_type_descriptor {
-    () => {
-        fn type_descriptor() -> $crate::__private::Cow<'static, $crate::type_info::TypeDescriptor> {
-            use $crate::__private::*;
-
-            let mut graph = TypeGraph::default();
-            let id = Self::build(&mut graph);
-            Cow::Owned(TypeDescriptor::__private_new(id, graph))
-        }
-    };
-}
-
 /// Reflected array types.
 pub mod array;
