+++ crates/mirror-mirror/src/std_impls/array.rs
@@ -27,8 +27,6 @@ where
         where
             T: Typed,
         {
-            fn_type_descriptor!();
-
             fn build(graph: &mut TypeGraph) -> NodeId {
                 graph.get_or_build_node_with::<Self, _>(|graph| ArrayNode::new::<Self, T, N>(graph))
             }
