+++ crates/mirror-mirror/src/struct_.rs
@@ -76,8 +76,6 @@ impl StructValue {
 impl Reflect for StructValue {
     fn type_descriptor(&self) -> Cow<'static, TypeDescriptor> {
         impl Typed for StructValue {
-            fn_type_descriptor!();
-
             fn build(graph: &mut TypeGraph) -> NodeId {
                 graph.get_or_build_node_with::<Self, _>(|graph| {
                     OpaqueNode::new::<Self>(Default::default(), graph)
