+++ crates/mirror-mirror/src/type_info/graph.rs
@@ -486,8 +486,6 @@ macro_rules! scalar_typed {
     ($($ty:ident)*) => {
         $(
             impl Typed for $ty {
-                fn_type_descriptor!();
-
                 fn build(graph: &mut TypeGraph) -> NodeId {
                     graph.get_or_build_node_with::<Self, _>(|_graph| ScalarNode::$ty)
                 }
