+++ crates/mirror-mirror/src/value.rs
@@ -174,8 +174,6 @@ macro_rules! for_each_variant {
 impl Reflect for Value {
     fn type_descriptor(&self) -> Cow<'static, TypeDescriptor> {
         impl Typed for Value {
-            fn_type_descriptor!();
-
             fn build(graph: &mut TypeGraph) -> NodeId {
                 graph.get_or_build_node_with::<Self, _>(|graph| {
                     OpaqueNode::new::<Self>(Default::default(), graph)
